Remove vanilla JS config (#3)
In `eslint-config-slapshot`, the base and vanilla JS configs don't include any linting rules besides code formatting. Since we will be using Prettier for code formatting instead, this exported config is no longer needed so this PR removes it and updates the README accordingly. For now, I left in the `overrides/javascript.js` file because it is used in the TS and React configs. In the items to update those, it will be deleted, or at least renamed.

Prettier is added and run, along with a GH action check for formatting to replace the linting check. The `package.json` has 2 new commands, `prettier-check` and `prettier-write`, to check the formatting and have Prettier re-write the files, respectively.

J=SLAP-2363
TEST=auto

See that the new GH check fails when code is not formatted with Prettier and passes after running `npm run prettier-write`.