Update TS config (#4)
Update the eslint config for TS to remove all rules for code formatting, as these will be handled by Prettier. We are no longer trying to use AirBnb as the base, so I moved the rules from `src/overrides/typescript.js` into `src/typescript.js`. For now, I left the former file as is because it is referenced in the React config. In the item to update that config, the "overrides" file will be removed.

J=SLAP-2364
TEST=manual

Use the config in Headless and run the CLI helper for `eslint-config-prettier` to verify that there are no conflicts between Prettier and this eslint config.