Update TSX config (#5)
Update the eslint config for Typescript React to remove all rules for code formatting, as these will be handled by Prettier. We are no longer trying to use AirBnb as the base, so I deleted the `overrides` folder and specified all the TSX rules, based on `eslint-config-slapshot`'s config for React, in `src/typescript-react.js`.

J=SLAP-2365
TEST=manual

Use the config in Headless-React and run the CLI helper for `eslint-config-prettier` to verify that there are no conflicts between Prettier and this eslint config.