Use eslint-config-prettier (#6)
Update the configs to use `eslint-config-prettier` to turn off rules that conflict with Prettier, instead of manually turning them off in the config.

J=SLAP-2393
TEST=manual

Use the eslint configs in Headless and Headless-React and see that they work as expected. Run the CLI helper for `eslint-config-prettier` to verify that there are no conflicts between Prettier and the configs.