Upgrade babel/traverse to version 7.24 to fix vulnerability (#9)
*ksearch: upgrade eslint-config to version 7.24

To resolve a vulnerability in babel/traverse <= 7.23, this PR
upgrades the version to 7.24. This was simply done via
npm install.

J=VULN-38972
TEST=manual

Made a clone of yext/studio locally, adjusted that package.json to refer to my local copy of this branch for @yext/eslint-config, and ran the autofix script. Output was identical to that of the default branch. Then introduced intentional errors (missing comma, etc) and ran the linting to see them reported.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>